+++ pkg/reconciliation/decommission_node_test.go
@@ -4,6 +4,7 @@
 package reconciliation
 
 import (
+	"fmt"
 	"net/http"
 	"strings"
 	"sync"
@@ -18,6 +19,8 @@ import (
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
+	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 )
 
 func TestRemoveDecommissionedPodFromZeroReplicaSts(t *testing.T) {
@@ -218,4 +221,113 @@ func TestRemoveResourcesWhenDone(t *testing.T) {
 	if r != result.RequeueSoon(5) {
 		t.Fatalf("expected result of blah but got %s", r)
 	}
+}
+
+func TestReconcileAllRacksDoesNotCleanUpDecommissioningNodeWhenMetadataRequestFails(t *testing.T) {
+	rc, _, cleanupMockScr := setupTest()
+	defer cleanupMockScr()
+
+	rc.Datacenter.Spec.Size = 1
+	rc.Datacenter.SetCondition(api.DatacenterCondition{
+		Status: corev1.ConditionTrue,
+		Type:   api.DatacenterScalingDown,
+	})
+
+	server := newFakeMgmtApiServer(t, http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		if r.URL.Path == "/api/v0/metadata/endpoints" {
+			w.WriteHeader(http.StatusInternalServerError)
+			return
+		}
+		http.NotFound(w, r)
+	}))
+	rc.NodeMgmtClient = server.client(rc.ReqLogger)
+
+	statefulSet, err := newStatefulSetForCassandraDatacenter(
+		nil,
+		"default",
+		rc.Datacenter,
+		2,
+		imageRegistry,
+	)
+	require.NoError(t, err)
+	statefulSet.Status.ObservedGeneration = statefulSet.Generation
+
+	podName := getStatefulSetPodNameForIdx(statefulSet, 1)
+	pvcName := fmt.Sprintf("%s-%s", PvcName, podName)
+	pod := &corev1.Pod{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      podName,
+			Namespace: rc.Datacenter.Namespace,
+			Labels: map[string]string{
+				api.ClusterLabel:    rc.Datacenter.Spec.ClusterName,
+				api.DatacenterLabel: rc.Datacenter.Name,
+				api.CassNodeState:   stateDecommissioning,
+				api.RackLabel:       "default",
+			},
+		},
+		Spec: corev1.PodSpec{
+			Containers: []corev1.Container{{Name: "cassandra"}},
+			Volumes: []corev1.Volume{{
+				Name: "server-data",
+				VolumeSource: corev1.VolumeSource{
+					PersistentVolumeClaim: &corev1.PersistentVolumeClaimVolumeSource{ClaimName: pvcName},
+				},
+			}},
+		},
+		Status: corev1.PodStatus{
+			ContainerStatuses: []corev1.ContainerStatus{{Name: "cassandra", Ready: true}},
+		},
+	}
+	server.attachToPod(t, pod)
+
+	pvc := &corev1.PersistentVolumeClaim{
+		ObjectMeta: metav1.ObjectMeta{Name: pvcName, Namespace: rc.Datacenter.Namespace},
+	}
+	remainingPod := pod.DeepCopy()
+	remainingPod.Name = getStatefulSetPodNameForIdx(statefulSet, 0)
+	remainingPod.Labels[api.CassNodeState] = stateStarted
+	remainingPVC := pvc.DeepCopy()
+	remainingPVC.Name = fmt.Sprintf("%s-%s", PvcName, remainingPod.Name)
+	remainingPod.Spec.Volumes[0].PersistentVolumeClaim.ClaimName = remainingPVC.Name
+	rc.Client = fake.NewClientBuilder().
+		WithScheme(setupScheme()).
+		WithStatusSubresource(rc.Datacenter, statefulSet).
+		WithRuntimeObjects(rc.Datacenter, statefulSet, pod, pvc, remainingPod, remainingPVC).
+		WithIndex(&corev1.Pod{}, podPVCClaimNameField, podPVCClaimNames).
+		Build()
+	rc.desiredRackInformation = []*RackInformation{{RackName: "default", NodeCount: 1, SeedCount: 1}}
+	rc.statefulSets = []*appsv1.StatefulSet{statefulSet}
+
+	_, err = rc.ReconcileAllRacks()
+	require.Error(t, err)
+	server.assertCallCount(t, "/api/v0/metadata/endpoints", 2)
+	server.assertCallCount(t, "/api/v0/ops/node/decommission", 0)
+
+	storedPVC := &corev1.PersistentVolumeClaim{}
+	require.NoError(t, rc.Client.Get(rc.Ctx, types.NamespacedName{
+		Name: pvcName, Namespace: rc.Datacenter.Namespace,
+	}, storedPVC))
+
+	storedStatefulSet := &appsv1.StatefulSet{}
+	require.NoError(t, rc.Client.Get(rc.Ctx, types.NamespacedName{
+		Name: statefulSet.Name, Namespace: statefulSet.Namespace,
+	}, storedStatefulSet))
+	require.Equal(t, int32(2), *storedStatefulSet.Spec.Replicas)
+}
+
+func TestDecommissionNodesRequiresMetadata(t *testing.T) {
+	rc, _, cleanupMockScr := setupTest()
+	defer cleanupMockScr()
+
+	rc.Datacenter.Spec.Size = 1
+	two := int32(2)
+	rc.statefulSets = []*appsv1.StatefulSet{{
+		Spec: appsv1.StatefulSetSpec{Replicas: &two},
+	}}
+	reconcileResult := rc.DecommissionNodes(httphelper.CassMetadataEndpoints{})
+	require.True(t, reconcileResult.Completed())
+	_, err := reconcileResult.Output()
+	require.Error(t, err)
+	require.NotEqual(t, corev1.ConditionTrue, rc.Datacenter.GetConditionStatus(api.DatacenterScalingDown))
+	require.Equal(t, int32(2), *rc.statefulSets[0].Spec.Replicas)
 }